requirements.txt

gradio
opencv-python
torch
sounddevice
numpy
scipy
pydub
gTTS
pygame
Pillow
torchaudio
torch
langchain-google-genai
langgraph
langchain-community
google-generativeai
langchain-huggingface  
langchain-tavily      
langchain-chroma
sentence-transformers
groq
requests
tavily-python
pywhatkit
pytesseract
pyzbar
python-dotenv